+++ pkgs/hooks/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.20.6-wip
+
+- Fix `testBuildHook` to emit output validation errors if there are any.
+
+## 0.20.5
+
+- Document `input.packageRoot` in more places.
+
 ## 0.20.4
 
 - Improved documentation on debugging and added a documentation index.